requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0

